+++ src/connectedk8s/HISTORY.rst
@@ -3,6 +3,12 @@
 Release History
 ===============
 
+0.2.9
+++++++
+* `az connectedk8s connect`: Added support for disabling auto upgrade of agents
+* `az connectedk8s update`: Added support for switching on/off the auto-upgrade
+* `az connectedk8s upgrade`: Added support for manual upgrading of agents
+
 0.2.8
 ++++++
 * Added checks for proxy and added disable-proxy